Add entity and field tags to schema data
This commit adds a 'fields' slot to the entity and fields objects used
in `schema.data`. This lets us tag entities and fields with their
intended use, or as required/optional, and whether they're managed by
the database or by contributors.

This commit also adds changes that make use of the new field to
color-code the entries in entity diagram and mark fields as required
and/or managed in the data dictionary.