Sync clinical guide and schema document
This commit syncs the contents of the schema document and the clinical
contributor's guide. This includes making field names consistent,
consistently ordering fields within tables (for consistency of
documentation only!), and carrying additions and deletions from one
document to the other where appropriate.

In future, it would be preferable to generate the clinical guide from
the data used to generate the schema document (if possible).