Add a ReSTructuredText target for exporting the schema
This commit adds a ReSTructuredText option for exporting the
schema. It generates a file of CSV tables with appropriate headings
containing the fields of each entity defined in `schema_data`. The
output is suitable for including in source for the SHARED
documentation, which should obviate the need for LaTeX output.