Fix pep8 violations in the schema data file
This is probably one of those "foolish consistency is the hobgoblin of
small mindes" moments, by the data file (with all its descriptions and
long lines) is now PEP8 compliant. The *point* of this is that now the
whole module is PEP8 compliant, and can be machine checked so that it
stays that way.