Add optional parens around drug lists
This commit adds parentheses around drug lists in the grammar for
regimens. Whereas before a list might be written

    100mg SOF TD + 200mg DAC TID 2 weeks

it could now be written

    (100mg SOF TD + 200mg DAC TID) 2 weeks

instead.

Technically, each paren is optional, so an imbalanced paren at the
beginning or end of a drug list would parse. But it would be silly to
write that.