Move brackets from DrugCombination to DoseList
This commit alters the regimen grammar so that DoseLists (compounds +
amounts combined by a `+`) may be surrounded by parentheses instead of
DrugCombinations (A list DoseLists and Frequencies joined by a
`&`). The resulting descriptions are hopefully a little more readable.

This commit also updates the test cases and standard regimens to
match.