Derive a SQLAlchemy DAO from the schema
This commit adds a Data Access object that holds SQLAlchemy table
references for each entity in the schema. These are generated
automatically from the schema.

The DAO uses the SQLAlchemy core API [0] instead of the SQLAlchemy ORM
API, on the assumption that we'll be making relatively few, but
relatively complicated queries.

[0]: https://docs.sqlalchemy.org/en/rel_1_1/core/index.html