Annotate submission scheme fields with the schema field they target
This commit adds an attribute to the submission scheme field object
which records the schema field that it will be loaded into (as a tuple
of entity name and field name). Submission scheme fields that don't
correspond to schema fields (e.g. sequence filenames, de-normalized
fields, etc.) have the `schema_path` attribute set to 'not
applicable'.

It also adds tests to check that all of the submission scheme fields
either point to real schema fields or are the exact string
'not applicable'.