Add a cannonical representation of datatypes
This commit adds an enum with members for each datatype in the schema,
and a classifier which takes a string describing a type and assigns it
to one of the categories.

This is NOT a parser that will go from a string description to a
concrete type (e.g. something that could parse a CSV file). But such a
parser could use the classifier to decide how to parse a type and
construct a loader (e.g. parse an ENUM type and verify that the data
being loaded is one of the possible values).