Replace the TreatmentData.duration_sch field
This commit replaces the TreatmentData.duration_sch field with a
number of fields on RegimenDrugInclusion. RegimenDrugInclusion can now
effectively represent regimens as parsed by the grammar in
`shared_schema.regimens`: there should be a RegimenDrugInclusion for
each medication in a regimen, and it will include the medication's
dose, frequency, and duration . Note that this means some information
will be repeated (e.g. the duration and frequency of two compounds in
the same pill). This representation is more suitable for storing in
the relational database; it will be converted from the tree structure
used by the grammar (which is more convenient for working with
notation) at load-time.