Change sequence filenames to sequence ids in submission schemes
This commit changes the way the submission schemes describe
sequences. Previouly, we expected a filename for each sequence, with
one sequence per file in the sequences directory. Now we expect a
sequence ID (stored in the sequence's title in the FASTA file) and
accept any number of files; the loading tool will search them for the
relevant sequence.