Change Medication from an entity to an enum type
This commit replaces the Medication table with a large enum type of
medication IDs. This changes the Regimen/Medication relationship from
a Many-to-many relation to a Many-to-one relation which
is *substantialy* simpler for domain experts to work with. The number
of medications is small enough and changes infrequently
enough (perhaps once a year, if the pace of research is very fast)
that making a schema change to add it to the ENUM type is not a
problem (especially with a SQLite back-end where the ENUM type is just
a check constraint anyways).