Fix a logic bug in regimens.cannonical
This commit fixes a logic bug in shared_schema.regimens.cannonical,
specifically in the code for loading standard regimens from the
DAO (to make the existing IDs of standard regimens avialable to code
that's loading new data). There's a method for loading the
RegimenDrugInclusion records of a Regimen from the database given the
regimen's ID; this does a JOIN on

    Reimgn.id = <the id passed as an arg>   AND
    Regimen.id = RegimenDrugInclusion.regimen_id

Due to an error in the query for returning a regimen's drug
inclusions, ALL of the drug inclusions were being loaded as a single
giant regimen. Specifically, the original code tried to combine the
two JOIN clauses with the `and` keyword, which (because of duck
typeing) actually returned just the second clause, which matches every
drug inclusion. The updated code uses `sqlalchemy.sql.and_` instead,
and works as expected.