Remove tex and html targets
This commit removes the generation code and build targets for TeX and
HTML forms of the schema. The function those targets served is now
being served (better) by the `shared-doc` module. They're being
removed here to reduce duplication.

This module does still export schema information, but does it in CSV
and RESructuredText format; `shared-doc` consumes that output to
generate tables and documentation.