Fix SqlAlchemy in-memory database problems
This commit makes a number of changes:

- Replace the `DAO.execute` method with `DAO.query` and `DAO.command`
  (for SQL that returns results or not, respectively).
- Update tests and internal methods to use the new API.
- Change the tests to use temporary files for databases instead of
  in-memory database.

This change was made because SqlAlchemy in-memory sqlite databases
behave differently than on-disk databases. Specifically, the API as
written *only* worked for in-memory databases (which is obviously not
what we want). This wasn't caught until much development had been
done using the existing API.