Disambiguate DAO.insert API
There appears to have been some ambiguity about what the propper
calling convention for the `DAO.insert` API is, and because it's
passing its arguments through to SQLAlchemy (which is very forgiving
about what it accepts) it appears to have gone unnoticed.

This commit adds a new method, `insert_many`, that accepts a list, and
adds tests to check that impropper use of the API will be caught.