Update dev dependencies
This commit changes the virtualenvironment creation process to:

- Update pip
- Not install dev tools (e.g. mypy and rope)

Updating pip is good security practice (it can warn if it gets too out
of date, and installing it from the standard library means it's not
going to get newer without updating Python).

I've started installing the dev tools globally (with `pip install
--user`) which has (so far) reduced some general hassle of dependency
conflicts within projects and configuring my IDE to find the tools (it
looks in the project VENV, but eventualy finds the relevant tools in
`~/.local/bin`).